Refactor autoCompleteInput into multiple files
