+++ javascript/src/utils/autoCompleteInput/domElements.js
@@ -1,0 +1,22 @@
+const notAvailableMessage = 'No hay opciones disponibles'
+
+export const unvailableItemTag = (searchText, addItems = false) => {
+  const li = document.createElement('li')
+
+  if (addItems && searchText && searchText.length > 0) {
+    li.append(`Agregar elemento: ${searchText}`)
+    li.dataset.value = searchText
+  } else {
+    li.append(notAvailableMessage)
+  }
+
+  return li
+}
+
+export const availableItemTag = item => {
+  const li = document.createElement('li')
+  const [name, value] = item
+  li.dataset.value = value
+  li.append(name)
+  return li
+}
